+++ be/src/runtime/load_stream_mgr.h
@@ -29,83 +29,19 @@
 #include "common/compiler_util.h" // IWYU pragma: keep
 #include "common/status.h"
 #include <runtime/sink_stream_mgr.h>
+#include <runtime/load_stream.h>
 
 namespace doris {
-
-class LoadStream : public StreamInputHandler {
-public:
-    LoadStream();
-    ~LoadStream();
-
-    int on_received_messages(StreamId id, butil::IOBuf* const messages[], size_t size) override;
-    void on_idle_timeout(StreamId id) override;
-    void on_closed(StreamId id) override;
-
-private:
-    void _handle_message(StreamId id, PStreamHeader hdr, TargetRowsetPtr rowset,
-                         TargetSegmentPtr segment, std::shared_ptr<butil::IOBuf> message);
-    void _parse_header(butil::IOBuf* const message, PStreamHeader& hdr);
-    Status _create_and_open_file(TargetSegmentPtr target_segment, std::string path);
-    Status _append_data(TargetSegmentPtr target_segment, std::shared_ptr<butil::IOBuf> message);
-    Status _close_file(TargetSegmentPtr target_segment);
-    void _report_status(StreamId stream, TargetRowsetPtr target_rowset, bool is_success,
-                        std::string error_msg);
-    uint64_t get_next_segmentid(TargetRowsetPtr target_rowset);
-    uint64_t get_next_segmentid(TargetRowsetPtr target_rowset, int64_t segmentid,
-                                int64_t backendid);
-    Status _build_rowset(TargetRowsetPtr target_rowset, const RowsetMetaPB& rowset_meta);
-
-private:
-    std::unique_ptr<ThreadPool> _workers;
-    // TODO: make it per load
-    std::map<TargetSegmentPtr, std::shared_ptr<io::LocalFileWriter>, TargetSegmentComparator>
-            _file_map;
-    std::mutex _file_map_lock;
-    // TODO: make it per load
-    std::map<TargetRowsetPtr, size_t, TargetRowsetComparator> _tablet_segment_next_id;
-    std::mutex _tablet_segment_next_id_lock;
-    std::map<TargetSegmentPtr, int64_t, TargetSegmentComparator> _tablet_segment_pos;
-    int64_t _current_id = 0;
-    // TODO: make it per load
-    std::map<TargetSegmentPtr, std::shared_ptr<ThreadPoolToken>, TargetSegmentComparator>
-            _segment_token_map; // accessed in single thread, safe
-    std::mutex _segment_token_map_lock;
-    std::map<TargetSegmentPtr, TargetSegmentPtr, TargetSegmentComparator>
-            _rawsegment_finalsegment_map;
-    std::mutex _rawsegment_finalsegment_map_lock;
-};
-
-using LoadStreamSharedPtr = std::shared_ptr<LoadStream>;
 
 class LoadStreamMgr {
 public:
     LoadStreamMgr() = default;
     ~LoadStreamMgr() = default;
 
-    LoadStreamSharedPtr find_or_create_load(PUniqueId loadid) {
-        std::string loadid_str = loadid.SerializeAsString();
-        std::lock_guard<std::mutex> l(_load_streams_lock);
-        if (_load_streams.find(loadid_str) == _load_streams.end()) {
-            _load_streams[loadid_str] = std::make_shared<LoadStream>();
-        }
-        return _load_streams[loadid_str];
-    }
-
-    void clear_load(PUniqueId loadid) {
-        std::string loadid_str = loadid.SerializeAsString();
-        _load_streams.erase(loadid_str);
-    }
-
-    Status bind_stream_to_load(LoadStreamSharedPtr loadstream, std::shared_ptr<StreamId> streamid) {
-        std::lock_guard<std::mutex> l(_load_stream_to_streamid_lock);
-        _load_stream_to_streamid[loadstream] = streamid;
-        return Status::OK();
-    }
-
-    Status unbind_stream_to_load(LoadStreamSharedPtr loadstream) {
-        CHECK(false); // not implemented
-        return Status::OK();
-    }
+    LoadStreamSharedPtr find_or_create_load(PUniqueId loadid);
+    void clear_load(PUniqueId loadid);
+    Status bind_stream_to_load(LoadStreamSharedPtr loadstream, std::shared_ptr<StreamId> streamid);
+    Status unbind_stream_to_load(LoadStreamSharedPtr loadstream);
 
 private:
     std::map<std::string, LoadStreamSharedPtr> _load_streams;
